Reuse seqn code for variable processing in seqn->satf
